Fixing the fucking bugs

+++ src/app/services/user.service.ts
@@ -671,7 +671,7 @@ export class UserService {
     }
   }
 
-  getExpertosDesac(route:string){
+  getExpertosDesac(route: string) {
     const token = localStorage.getItem('token');
     console.log(token);
 
@@ -694,7 +694,7 @@ export class UserService {
     }
   }
 
-//Traer datos de un usuario por su id 
+  //Traer datos de un usuario por su id 
   getUserOnlyId(id: number): Observable<any> {
     // Obtener el token del local storage
     console.log(id);
@@ -718,38 +718,46 @@ export class UserService {
 
   }
 
-//Actualizar informacion del usuario por id 
+  // Acuatlizar datos de usuario por id
   updateUserById(id: number, formData: FormData): Observable<any> {
-    // Obtener el token del local storage
-    console.log(id);
-  activarExpert(id: number): Observable<any> {
     // Obtener el token del local storage
     const token = localStorage.getItem('token');
     console.log(token);
-
     // Verificar si el usuario está autenticado
     if (this.isAuth() && token) {
-
       console.log('Entro al if');
       // Configurar las cabeceras con el token de autenticación
       const headers = new HttpHeaders({
         'Authorization': `Bearer ${token}`
       });
+      console.log('token', headers);
       // Realizar la solicitud a la API utilizando el token en las cabeceras
       return this.http.put(this.url + `/api/actualizarusuario/${id}`, formData, { headers });
     } else {
-      return new Observable(); // Puedes también retornar throwError o un observable vacío, según tu necesidad.
+      return new Observable();
     }
   }
 
+  //Activar experto
+  activarExpert(id: number): Observable<any> {
+    // Obtener el token del local storage
+    const token = localStorage.getItem('token');
+    console.log(token);
+    // Verificar si el usuario está autenticado
+    if (this.isAuth() && token) {
+      console.log('Entro al if');
+      // Configurar las cabeceras con el token de autenticación
+      const headers = new HttpHeaders({
+        'Authorization': `Bearer ${token}`
+      });
+      console.log('token', headers);
+      // Realizar la solicitud a la API utilizando el token en las cabeceras
       return this.http.delete(this.url + `/api/activaruser/${id}`, { headers });
     } else {
-      // Si el usuario no está autenticado o no hay token, redirigir a la página de inicio de sesión u otra página apropiada.
-      // Por ejemplo, puedes utilizar un guard para proteger la ruta y redirigir en caso de que el usuario no esté autenticado.
-      // Aquí retornamos un observable vacío, pero puedes manejar el redireccionamiento según tu lógica.
-      return new Observable(); // Puedes también retornar throwError o un observable vacío, según tu necesidad.
+      return new Observable();
     }
   }
+
 
 }
